refactor(mediaplayer): unify audio into one BasisMediaPlayerAudio component
Collapse the audio sink to a single component. BasisMediaPlayerAudio handles
both stereo and multichannel by routing every output through the Outputs array,
each AudioSource tagged with a BasisMediaAudioChannel (a single decoded channel,
or a stereo downmix). Stereo is one Output set to Stereo; surround is one Output
per channel.

- Remove BasisMediaPlayerMultiChannelAudio; its implementation becomes
  BasisMediaPlayerAudio, retaining the original component GUID so existing
  content loads the unified component rather than erroring. Content built
  against the old single-source component must relink its AudioSource into
  Outputs.
- Remove the BasisAudioRouting enum and the dual-sink routing in
  BasisMediaPlayer; there is one sink now.
- Remove the managed BasisAudioFrame path. OnAudioFrame had no producer (the
  synthetic source stubbed it; the native engine delivers audio via the PCM ring
  and owns A/V sync via PositionUs), so its frame queue and audio-derived clock
  never ran on any stream. Drop BasisAudioFrame and IBasisFrameSource.OnAudioFrame.
- MediaPlayerStreaming prefab uses the Outputs path (one source set to Stereo);
  MediaPlayerMultiChannelStreaming repointed to the unified component and centred
  on the Front Centre speaker.
- ConsumedSampleCount counts sample-frames (not interleaved floats) so it is
  consistent across mono and stereo primary outputs; Sample Rate / Channel Count
  inspector fields are read-only (decoder-driven).
- Dedupe the audio rebuild against the last announced format so an aborted
  rebuild (e.g. no Outputs wired) is not retried every frame.
- Update inspectors, debug window, diagnostics, and README.
