--- Basis/Packages/com.basis.mediaplayer/Editor/BasisMediaPlayerMultiChannelAudioInspector.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: bc40a8d4ccdf4dc4a6c61fcaa46ced84
