+++ Basis/Packages/com.basis.mediaplayer/Editor/StyleSheets/MediaPlayerAudioSDK.uxml
@@ -1,39 +1,28 @@
-<engine:UXML xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:engine="UnityEngine.UIElements" xmlns:editor="UnityEditor.UIElements" noNamespaceSchemaLocation="../../UIElementsSchema/UIElements.xsd" editor-extension-mode="False">
+<engine:UXML xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:engine="UnityEngine.UIElements" xmlns:editor="UnityEditor.UIElements" noNamespaceSchemaLocation="../../UIElementsSchema/UIElements.xsd" editor-extension-mode="True">
     <engine:VisualElement class="bvp-root">
         <engine:VisualElement name="HeaderCard" class="bvp-card bvp-header-card">
             <engine:Label text="Basis Media Audio" name="HeaderTitle" selectable="false" class="bvp-header-title"/>
-            <engine:Label text="Streaming AudioClip sink + IBasisMediaClockSource" name="HeaderSubtitle" selectable="false" class="bvp-header-subtitle"/>
+            <engine:Label text="One AudioSource per channel, or a stereo downmix" name="HeaderSubtitle" selectable="false" class="bvp-header-subtitle"/>
         </engine:VisualElement>
 
         <engine:Foldout text="Output" name="OutputFoldout" value="true" class="bvp-foldout">
-            <editor:ObjectField label="Target Audio Source" name="TargetAudioSourceField" type="UnityEngine.AudioSource, UnityEngine.AudioModule" allow-scene-objects="true"/>
-            <engine:Toggle label="Assign Clip On Awake" name="AssignClipOnAwakeField"/>
-            <engine:TextField label="Clip Name" name="ClipNameField"/>
+            <editor:PropertyField name="OutputsField" label="Outputs" tooltip="AudioSources that play content audio. Each needs a BasisMediaAudioChannel selecting its channel(s) — a single channel, or a stereo downmix. Used for both stereo and surround setups."/>
         </engine:Foldout>
 
         <engine:Foldout text="Format" name="FormatFoldout" value="false" class="bvp-foldout">
-            <engine:IntegerField label="Sample Rate" name="SampleRateField" tooltip="Auto-updated from the first incoming audio frame."/>
-            <engine:SliderInt label="Channel Count" name="ChannelCountField" low-value="1" high-value="2" show-input-field="true"/>
+            <engine:IntegerField label="Sample Rate" name="SampleRateField" enabled="false" tooltip="Read-only; auto-updated from the decoder."/>
+            <engine:SliderInt label="Channel Count" name="ChannelCountField" low-value="1" high-value="8" show-input-field="true" enabled="false" tooltip="Read-only; auto-updated from the decoder."/>
         </engine:Foldout>
 
         <engine:Foldout text="Buffering" name="BufferingFoldout" value="false" class="bvp-foldout">
-            <engine:FloatField label="Clip Length (s)" name="ClipLengthField" tooltip="Larger = coarser PCM callbacks; ~1.0 keeps callbacks around 80 ms."/>
-            <engine:IntegerField label="Max Queued Frames" name="MaxQueuedField" tooltip="0 = unbounded."/>
-            <engine:Toggle label="Drop Oldest On Overflow" name="DropOldestField"/>
-            <engine:IntegerField label="Rebuffer Frames" name="RebufferField" tooltip="After underrun, stay silent until this many frames have queued."/>
+            <engine:FloatField label="Clip Length (s)" name="ClipLengthField" tooltip="Length of each streaming AudioClip and the depth of the broadcast window. Larger = steadier but adds latency."/>
         </engine:Foldout>
 
         <engine:Foldout text="Playback" name="PlaybackFoldout" value="true" class="bvp-foldout">
             <engine:Toggle label="Auto Play On Enable" name="AutoPlayField"/>
             <engine:Toggle label="Stop On Disable" name="StopOnDisableField"/>
-            <engine:Toggle label="Clear Queue On Enable" name="ClearQueueOnEnableField"/>
             <engine:Slider label="Volume Gain" name="VolumeGainField" low-value="0" high-value="2" show-input-field="true" tooltip="Sample-domain multiplier applied after decode (hard-capped at 2.0 at runtime)."/>
             <engine:Toggle label="Mute" name="MuteField"/>
         </engine:Foldout>
-
-        <engine:VisualElement name="ActionsCard" class="bvp-card bvp-actions-card">
-            <engine:Button text="Clear Queue" name="ClearQueueButton" class="bvp-primary-button"/>
-            <engine:Button text="Reset Sync Anchor" name="ResetAnchorButton" class="bvp-secondary-button"/>
-        </engine:VisualElement>
     </engine:VisualElement>
 </engine:UXML>
