--- Basis/Packages/com.basis.mediaplayer/Editor/StyleSheets/MediaPlayerMultiChannelSDK.uxml.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 723983a53f2e4b83a330d96368c2bc6c
-ScriptedImporter:
-  internalIDToNameTable: []
-  externalObjects: {}
-  serializedVersion: 2
-  userData:
-  assetBundleName:
-  assetBundleVariant:
-  script: {fileID: 13804, guid: 0000000000000000e000000000000000, type: 0}
