+++ Basis/Packages/com.basis.mediaplayer/Prefabs/MediaPlayerMultiChannelStreaming.prefab
@@ -131,7 +131,6 @@ MonoBehaviour:
   Loop: 0
   LoopRestartDelaySeconds: 0
   PresentationOffsetUs: 0
-  AudioRouting: 1
   Volume: 1
   Mute: 0
   PlaybackRate: 1
@@ -177,9 +176,9 @@ MonoBehaviour:
   m_GameObject: {fileID: 692773274864663004}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 7f3e949b42bd414393e9addb2a0714c9, type: 3}
+  m_Script: {fileID: 11500000, guid: b8f0e4832d2e9ba4bbf3db4b90cc1c7c, type: 3}
   m_Name: 
-  m_EditorClassIdentifier: BasisMediaPlayer::BasisMediaPlayerMultiChannelAudio
+  m_EditorClassIdentifier: BasisMediaPlayer::BasisMediaPlayerAudio
   Outputs:
   - {fileID: 5816047855775780060}
   - {fileID: 1169068719169778331}
@@ -482,7 +481,7 @@ Transform:
   m_GameObject: {fileID: 1006988167474170037}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 4, y: 0, z: -4}
+  m_LocalPosition: {x: 4, y: 0, z: -8}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -625,7 +624,7 @@ Transform:
   m_GameObject: {fileID: 2565437961886034552}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 4, y: 0, z: 3}
+  m_LocalPosition: {x: 4, y: 0, z: -1}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -768,7 +767,7 @@ Transform:
   m_GameObject: {fileID: 4067286564769727513}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: -4, y: 0, z: -4}
+  m_LocalPosition: {x: -4, y: 0, z: -8}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -911,7 +910,7 @@ Transform:
   m_GameObject: {fileID: 6086984266641354651}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: -4, y: 0, z: -0.5}
+  m_LocalPosition: {x: -4, y: 0, z: -4.5}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -1054,7 +1053,7 @@ Transform:
   m_GameObject: {fileID: 6556824477466809312}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 4, y: 0, z: -0.5}
+  m_LocalPosition: {x: 4, y: 0, z: -4.5}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -1197,7 +1196,7 @@ Transform:
   m_GameObject: {fileID: 6895909008755713970}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: 0, z: 4}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -1340,7 +1339,7 @@ Transform:
   m_GameObject: {fileID: 7106499004218466471}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: -4, y: 0, z: 3}
+  m_LocalPosition: {x: -4, y: 0, z: -1}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -1483,7 +1482,7 @@ Transform:
   m_GameObject: {fileID: 7370652399459993111}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: -2, z: 3}
+  m_LocalPosition: {x: 0, y: -2, z: -1}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
