--- Basis/Packages/com.basis.mediaplayer/Runtime/Audio/BasisMediaPlayerMultiChannelAudio.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 7f3e949b42bd414393e9addb2a0714c9
