--- Basis/Packages/com.basis.mediaplayer/Runtime/Core/BasisAudioFrame.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: d01d917ada8da9e46873745f92ea5476
