--- Basis/Packages/com.basis.mediaplayer/Runtime/Sources/BasisAudioRouting.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: b223eb1f5029bcc48be85b6cb0e570d2
